requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anthropic
anyio
asyncio
attrs
beautifulsoup4
bs4
cacherator
cachetools
certifi
charset-normalizer
click
colorama
distro
docorator
feedparser
filelock
frozenlist
ghostscraper
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
gspread
h11
html2text
htmldocx
httpcore
httplib2
httpx
idna
jiter
joblib
logorator
lxml
lxml_html_clean
Markdown
multidict
newspaper4k
nltk
numpy
oauthlib
openai
pandas
pillow
playwright
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyee
pyparsing
python-dateutil
python-docx
python-slugify
pytz
PyYAML
regex
requests
requests-file
requests-oauthlib
rsa
searcherator
sgmllib3k
six
smartllm
smartspread
sniffio
soupsieve
text-unidecode
tldextract
toml_i18n
tqdm
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
yarl
